+++ packages/react-bootstrap/src/CheckboxWidget/CheckboxWidget.tsx
@@ -50,7 +50,7 @@ export default function CheckboxWidget<
   const description = options.description || schema.description;
   return (
     <Form.Group className={disabled || readonly ? 'disabled' : ''} aria-describedby={ariaDescribedByIds<T>(id)}>
-      {!hideLabel && !!description && (
+      {!hideLabel && description && (
         <DescriptionFieldTemplate
           id={descriptionId<T>(id)}
           description={description}